sensor/requirements.txt
kafka-python
scapy-python3
